Moved model to a new folder.
