+++ auth/src/main/java/api/v1/UserIdentityServlet.java
@@ -3,6 +3,9 @@ package api.v1;
 import exception.BadDataException;
 import exception.FailedCRUDOperationException;
 import exception.UnsupportedMediaTypeException;
+import model.response.IdentityPayload;
+import model.response.ResponseBody;
+import model.uid.UserIdentity;
 import service.UserIdentityService;
 
 import javax.servlet.ServletException;
@@ -10,11 +13,43 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
+import java.io.PrintWriter;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 public class UserIdentityServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        response.setContentType("application/json");
+        PrintWriter output = response.getWriter();
 
+        String message = "";
+        List<IdentityPayload> payloads = new ArrayList<>();
+
+        try {
+            List<UserIdentity> userIdentities = UserIdentityService.getAll();
+            payloads = toPayloadList(userIdentities);
+            message = "Get success.";
+            response.setStatus(HttpServletResponse.SC_OK);
+
+        } catch (SQLException e) {
+            message = (e.getMessage());
+            response.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+
+        } finally {
+            ResponseBody<IdentityPayload> responseBody = new ResponseBody<>(message, payloads);
+            output.write(responseBody.toJSON());
+        }
+    }
+
+    private List<IdentityPayload> toPayloadList(List<UserIdentity> userIdentities) {
+        List<IdentityPayload> payloads = new ArrayList<>();
+        for (UserIdentity uid :
+                userIdentities) {
+            payloads.add(new IdentityPayload(uid));
+        }
+        return payloads;
     }
 
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
