+++ auth/src/test/java/UserIdentityTest.java
@@ -1,6 +1,6 @@
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import model.UserIdentity;
+import model.uid.UserIdentity;
 
 public class UserIdentityTest {
     public static void main(String[] args) throws JsonProcessingException {
